% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmh_hydraulic_coef_data.R
\docType{data}
\name{nmh_hydraulic_coef_data}
\alias{nmh_hydraulic_coef_data}
\title{hydraulic coefficient data for NEON sites}
\format{
An object of class \code{data.frame} with 24 rows and 18 columns.
}
\usage{
nmh_hydraulic_coef_data
}
\description{
hydraulic coefficient data for NEON sites
}
\author{
Nick Marzolf \email{nickmarzolf@gmail.com}
}
\keyword{data}
