% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmh_q_eval.R
\docType{data}
\name{nmh_q_eval}
\alias{nmh_q_eval}
\title{Rhea et al. NEON discharge evaulation data}
\format{
An object of class \code{data.frame} with 804 rows and 19 columns.
}
\usage{
nmh_q_eval
}
\description{
Rhea et al. NEON discharge evaulation data
}
\author{
Spencer Rhea \email{spencerrhea@gmail.com}
}
\keyword{data}
